GoogleTranslates

Documentation for GoogleTranslates.

GoogleTranslates.excel_colname

GoogleTranslates.getcodes

GoogleTranslates.mds2temp

GoogleTranslates.read_xlsx

GoogleTranslates.splitMDtext

GoogleTranslates.temp2mds

GoogleTranslates.write_xlsx

GoogleTranslates.write_xlsx

GoogleTranslates.excel_colname — Method

excel_colind(k::Int)

Convert the column index to a column name.

GoogleTranslates.getcodes — Method

getcodes(txt::AbstractString)

Get the code blocks from txt.

GoogleTranslates.mds2temp — Method

mds2temp( filesin::AbstractVector{<:AbstractString}
        , excel::AbstractString="temp.xlsx"
        , info::AbstractString="temp.txt")

Convert markdown files to temp files.

GoogleTranslates.read_xlsx — Method

read_xlsx(filename::AbstractString)

Read the first sheet of the Excel file into a matrix of strings.

GoogleTranslates.splitMDtext — Method

splitMDtext(text::AbstractString)

Split a markdown text into a list of markdown blocks.

GoogleTranslates.temp2mds — Method

temp2mds( files::AbstractVector{<:AbstractString}
        , info::AbstractString="temp.txt"
        , excel::AbstractString="temp.xlsx")

Recover markdown files from temp files.

GoogleTranslates.write_xlsx — Method

write_xlsx(filename::AbstractString, mat::AbstractMatrix)

Write the matrix to an Excel file.

GoogleTranslates.write_xlsx — Method

write_xlsx(filename::AbstractString, vector::AbstractVector)

Write the vector to the first column of an Excel file.